+++ products/experiments/backend/test/test_presentation_api.py
@@ -5930,7 +5930,7 @@ class TestExperimentParametersFieldMutation(APILicensedTest):
     """
 
     def test_to_internal_value_does_not_mutate_input_dict(self):
-        from products.experiments.backend.presentation.experiments import ExperimentParametersField
+        from products.experiments.backend.presentation.serializers import ExperimentParametersField
 
         input_dict = {
             "feature_flag_variants": [
@@ -5950,7 +5950,7 @@ class TestExperimentParametersFieldMutation(APILicensedTest):
         }
 
     def test_to_representation_does_not_mutate_stored_value(self):
-        from products.experiments.backend.presentation.experiments import ExperimentParametersField
+        from products.experiments.backend.presentation.serializers import ExperimentParametersField
 
         stored_value = {
             "feature_flag_variants": [
